refactor(Function): 함수를 저장하는 파일을 만들어 분리
